backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
firebase-admin==6.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
